#33 temporarily move SmtpTransport where it was to ease the PR review.
